--- src/main/resources/assets/css/page/form.css
@@ -1,94 +1,0 @@
-#doc-ipt-pwd-1{
-	background-color: #FFFFFF;
-    border: 1px solid #E3E3E3;
-    border-radius: 4px;
-    color: #565656;
-    padding: 7px 12px;
-    height: 38px;
-    max-width: 100%;
-    -webkit-box-shadow: none;
-    box-shadow: none;
-    -webkit-transition: all 300ms linear;
-    -moz-transition: all 300ms linear;
-    -o-transition: all 300ms linear;
-    -ms-transition: all 300ms linear;
-    transition: all 300ms linear;
-    width: 83.33333333%;
-}
-
-
-
-#doc-ta-1,#doc-ds-ipt-1,#doc-select-1{
-    border: 1px solid #E3E3E3;
-    border-radius: 4px;
-    color: #565656;
-    padding: 7px 12px;
-    max-width: 100%;
-    -webkit-box-shadow: none;
-    box-shadow: none;
-    -webkit-transition: all 300ms linear;
-    -moz-transition: all 300ms linear;
-    -o-transition: all 300ms linear;
-    -ms-transition: all 300ms linear;
-    transition: all 300ms linear;
-    width: 83.33333333%;
-}
-
-#rightPa{
-	right: 25px;
-}
-
-.help-block {
-    display: block;
-    margin-top: 5px;
-    margin-bottom: 10px;
-    color: rgb(115, 115, 115);
-}
-
-@media screen and (max-width: 641px){
-	#doc-ipt-pwd-1,#doc-ta-1,#doc-ds-ipt-1,#doc-select-1{
-		width: 100%;
-	}
-	
-
-}
-
-
-
-/*----------------------登录-----------------------*/
-.account-pages {
-    height: 100%;
-    width: 100%;
-    background:url(../../img/bg1.jpg) center !important;
-    background-size: cover;
-}
-.wrapper-page {
-    margin: 5% auto;
-    position: relative;
-    width: 420px;
-}
-@media (max-width: 768px){
-.wrapper-page {
-    width: 90%;
-}
-}
-.text-center {
-    text-align: center;
-}
-.wrapper-page .logo {
-    font-size: 42px;
-}
-.logo {
-    color: #188ae2 !important;
-    font-size: 32px;
-    font-family: 'Rancho', cursive;
-    text-decoration: none;
-}
-.panel-body {
-    padding: 15px;
-}
-.remeber{
-	width: 15px;
-	height: 15px;
-}
-
