--- src/main/resources/assets/css/page/typography.css
@@ -1,4 +1,0 @@
-.card-box .am-nav > li > a:hover, .am-nav > li > a:focus {
-    text-decoration: none;
-    background-color: none;
-}
