+++ src/main/resources/static/assets/js/charts/lineChart.js
@@ -323,7 +323,8 @@ option = {
 	
 	function randomData() {
     now = new Date(+now + oneDay);
-    value = value + Math.random() * 21 - 10;
+
+    value = value + Math.random() * 8;
     return {
         name: now.toString(),
         value: [
@@ -334,16 +335,16 @@ option = {
 }
 
 var data = [];
-var now = +new Date(1997, 9, 3);
+var now = new Date()-24*3600*1000*30;
 var oneDay = 24 * 3600 * 1000;
-var value = Math.random() * 1000;
-for (var i = 0; i < 1000; i++) {
+var value = Math.random() * 5;
+for (var i = 0; i < 30; i++) {
     data.push(randomData());
 }
 
 option = {
     title: {
-        text: '动态数据 + 时间坐标轴'
+        text: '访问量'
     },
     tooltip: {
         trigger: 'axis',
@@ -378,19 +379,19 @@ option = {
     }]
 };
 
-	timeTicket = setInterval(function () {
-	
-	    for (var i = 0; i < 5; i++) {
-	        data.shift();
-	        data.push(randomData());
-	    }
-	
-	    trends.setOption({
-	        series: [{
-	            data: data
-	        }]
-	    });
-	}, 1000);
+	// timeTicket = setInterval(function () {
+    //
+	//     for (var i = 0; i < 5; i++) {
+	//         data.shift();
+	//         data.push(randomData());
+	//     }
+    //
+	//     trends.setOption({
+	//         series: [{
+	//             data: data
+	//         }]
+	//     });
+	// }, 1000);
 
 trends.setOption(option);
 })();
@@ -410,6 +411,85 @@ trends.setOption(option);
 
 
 
+//动态数据+时间坐标轴
+(function(){
+
+    var trends = echarts.init(document.getElementById("trends_number"));
+
+    function randomData() {
+        now = new Date(+now + oneDay);
+
+        value = value + Math.random() * 3;
+        return {
+            name: now.toString(),
+            value: [
+                [now.getFullYear(), now.getMonth() + 1, now.getDate()].join('/'),
+                Math.round(value)
+            ]
+        }
+    }
+
+    var data = [];
+    var now = new Date()-24*3600*1000*30;
+    var oneDay = 24 * 3600 * 1000;
+    var value = Math.random() * 2;
+    for (var i = 0; i < 30; i++) {
+        data.push(randomData());
+    }
+
+    option = {
+        title: {
+            text: '注册用户'
+        },
+        tooltip: {
+            trigger: 'axis',
+            formatter: function (params) {
+                params = params[0];
+                var date = new Date(params.name);
+                return date.getDate() + '/' + (date.getMonth() + 1) + '/' + date.getFullYear() + ' : ' + params.value[1];
+            },
+            axisPointer: {
+                animation: false
+            }
+        },
+        xAxis: {
+            type: 'time',
+            splitLine: {
+                show: false
+            }
+        },
+        yAxis: {
+            type: 'value',
+            boundaryGap: [0, '100%'],
+            splitLine: {
+                show: false
+            }
+        },
+        series: [{
+            name: '模拟数据',
+            type: 'line',
+            showSymbol: false,
+            hoverAnimation: false,
+            data: data
+        }]
+    };
+
+    // timeTicket = setInterval(function () {
+    //
+    //     for (var i = 0; i < 5; i++) {
+    //         data.shift();
+    //         data.push(randomData());
+    //     }
+    //
+    //     trends.setOption({
+    //         series: [{
+    //             data: data
+    //         }]
+    //     });
+    // }, 1000);
+
+    trends.setOption(option);
+})();
